refactor: split 7 files >500 lines into focused modules (Fowler Extract Module)
Applied Martin Fowler's Extract Module pattern to reduce cognitive complexity:

- x86_avx2.rs (556L) → x86_avx2/{mod,dot,l2}.rs (327+224+13)
- guardrails.rs (520L) → guardrails/{mod,limits,context,resilience}.rs (87+177+124+173)
- grouped.rs (522L) → grouped.rs + having.rs (235+305)
- search.rs (519L) → search.rs + brute_force.rs (328+153)
- edge_concurrent.rs (509L) → edge_concurrent/{mod,query}.rs (302+151)
- node.rs (506L) → node/{mod,growth}.rs (419+98)
- gpu_backend.rs (501L) → gpu_backend.rs + shaders.rs (364+114)

Also fixed:
- database_tests.rs: added missing 'use crate::point::Point' import
- search.rs: added missing Hamming/Jaccard variants in compute_distance

All files now < 500 lines. cargo check + cargo test --no-run pass with 0 errors.
