+++ crates/velesdb-core/src/collection/graph/cart/node/mod.rs
@@ -1,13 +1,16 @@
-//! C-ART node types and operations.
+//! C-ART node implementation.
 //!
 //! Implements the internal node variants (Node4, Node16, Node48, Node256, Leaf)
-//! with search, insert, remove, and growth operations.
+//! with search, insert, remove, and collect operations.
+//! Growth operations (Node4→Node16→Node48→Node256) are in `growth.rs`.
 
 // SAFETY: Numeric casts in C-ART node operations are intentional:
 // - usize->u8 for child indices: C-ART nodes have max 256 children, indices fit in u8
 // - Node types enforce size limits (Node4=4, Node16=16, Node48=48, Node256=256)
 // - All index values are validated against node capacity before casting
 #![allow(clippy::cast_possible_truncation)]
+
+mod growth;
 
 /// Node variants for the Compressed Adaptive Radix Tree.
 // Reason: Node256 is larger than other variants by design for high-degree vertices
@@ -60,7 +63,7 @@ impl CARTNode {
         }
     }
 
-    /// Checks if this node is empty.
+    /// Returns true if this node is empty.
     pub(crate) fn is_empty(&self) -> bool {
         match self {
             Self::Leaf { entries, .. } => entries.is_empty(),
@@ -79,6 +82,7 @@ impl CARTNode {
                 num_children,
                 keys,
                 children,
+                ..
             } => {
                 if key.is_empty() {
                     return false;
@@ -97,12 +101,12 @@ impl CARTNode {
                 num_children,
                 keys,
                 children,
+                ..
             } => {
                 if key.is_empty() {
                     return false;
                 }
                 let byte = key[0];
-                // Binary search for SIMD-friendly access
                 let slice = &keys[..*num_children as usize];
                 if let Ok(idx) = slice.binary_search(&byte) {
                     if let Some(child) = &children[idx] {
@@ -128,11 +132,12 @@ impl CARTNode {
                 if key.is_empty() {
                     return false;
                 }
-                let byte = key[0];
-                if let Some(child) = &children[byte as usize] {
-                    return child.search(&key[1..], value);
+                let byte = key[0] as usize;
+                if let Some(child) = &children[byte] {
+                    child.search(&key[1..], value)
+                } else {
+                    false
                 }
-                false
             }
         }
     }
@@ -142,11 +147,9 @@ impl CARTNode {
     pub(crate) fn insert(&mut self, key: &[u8], value: u64) -> bool {
         match self {
             Self::Leaf { entries, .. } => {
-                // Binary search for insertion point
                 match entries.binary_search(&value) {
                     Ok(_) => false, // Already exists
                     Err(pos) => {
-                        // Leaf stores sorted local bucket; internal-node growth handles branching.
                         entries.insert(pos, value);
                         true
                     }
@@ -162,7 +165,6 @@ impl CARTNode {
                 }
                 let byte = key[0];
 
-                // Check if key exists
                 for i in 0..*num_children as usize {
                     if keys[i] == byte {
                         if let Some(child) = &mut children[i] {
@@ -171,7 +173,6 @@ impl CARTNode {
                     }
                 }
 
-                // Key doesn't exist, add new child
                 if (*num_children as usize) < 4 {
                     let idx = *num_children as usize;
                     keys[idx] = byte;
@@ -179,7 +180,6 @@ impl CARTNode {
                     *num_children += 1;
                     true
                 } else {
-                    // Node is full, need to grow to Node16
                     *self = self.grow_to_node16();
                     self.insert(key, value)
                 }
@@ -194,7 +194,6 @@ impl CARTNode {
                 }
                 let byte = key[0];
 
-                // Binary search for key
                 let slice = &keys[..*num_children as usize];
                 match slice.binary_search(&byte) {
                     Ok(idx) => {
@@ -206,7 +205,6 @@ impl CARTNode {
                     }
                     Err(pos) => {
                         if (*num_children as usize) < 16 {
-                            // Shift elements to maintain sorted order
                             let n = *num_children as usize;
                             for i in (pos..n).rev() {
                                 keys[i + 1] = keys[i];
@@ -217,7 +215,6 @@ impl CARTNode {
                             *num_children += 1;
                             true
                         } else {
-                            // Grow to Node48
                             *self = self.grow_to_node48();
                             self.insert(key, value)
                         }
@@ -236,13 +233,11 @@ impl CARTNode {
                 let slot = keys[byte as usize];
 
                 if slot != 255 {
-                    // Key exists, recurse
                     if let Some(child) = &mut children[slot as usize] {
                         return child.insert(&key[1..], value);
                     }
                 }
 
-                // Key doesn't exist, add new child
                 if (*num_children as usize) < 48 {
                     let new_slot = *num_children;
                     keys[byte as usize] = new_slot;
@@ -250,7 +245,6 @@ impl CARTNode {
                     *num_children += 1;
                     true
                 } else {
-                    // Grow to Node256
                     *self = self.grow_to_node256();
                     self.insert(key, value)
                 }
@@ -301,9 +295,9 @@ impl CARTNode {
                         if let Some(child) = &mut children[i] {
                             let removed = child.remove(&key[1..], value);
                             if removed && child.is_empty() {
-                                // Remove empty child
+                                children[i] = None;
                                 let n = *num_children as usize;
-                                for j in i..(n - 1) {
+                                for j in i..n - 1 {
                                     keys[j] = keys[j + 1];
                                     children[j] = children[j + 1].take();
                                 }
@@ -329,8 +323,9 @@ impl CARTNode {
                     if let Some(child) = &mut children[idx] {
                         let removed = child.remove(&key[1..], value);
                         if removed && child.is_empty() {
+                            children[idx] = None;
                             let n = *num_children as usize;
-                            for j in idx..(n - 1) {
+                            for j in idx..n - 1 {
                                 keys[j] = keys[j + 1];
                                 children[j] = children[j + 1].take();
                             }
@@ -419,88 +414,6 @@ impl CARTNode {
                     child.collect_all(result);
                 }
             }
-        }
-    }
-
-    /// Grows Node4 to Node16.
-    fn grow_to_node16(&self) -> Self {
-        match self {
-            Self::Node4 {
-                num_children,
-                keys,
-                children,
-            } => {
-                let mut new_keys = [0u8; 16];
-                let mut new_children: [Option<Box<CARTNode>>; 16] = Default::default();
-
-                // Copy and sort
-                let n = *num_children as usize;
-                let mut indices: Vec<usize> = (0..n).collect();
-                indices.sort_by_key(|&i| keys[i]);
-
-                for (new_idx, &old_idx) in indices.iter().enumerate() {
-                    new_keys[new_idx] = keys[old_idx];
-                    new_children[new_idx].clone_from(&children[old_idx]);
-                }
-
-                Self::Node16 {
-                    num_children: *num_children,
-                    keys: new_keys,
-                    children: new_children,
-                }
-            }
-            _ => self.clone(),
-        }
-    }
-
-    /// Grows Node16 to Node48.
-    fn grow_to_node48(&self) -> Self {
-        match self {
-            Self::Node16 {
-                num_children,
-                keys,
-                children,
-            } => {
-                let mut new_keys = [255u8; 256];
-                let mut new_children: [Option<Box<CARTNode>>; 48] = std::array::from_fn(|_| None);
-
-                for i in 0..*num_children as usize {
-                    new_keys[keys[i] as usize] = i as u8;
-                    new_children[i].clone_from(&children[i]);
-                }
-
-                Self::Node48 {
-                    num_children: *num_children,
-                    keys: new_keys,
-                    children: new_children,
-                }
-            }
-            _ => self.clone(),
-        }
-    }
-
-    /// Grows Node48 to Node256.
-    fn grow_to_node256(&self) -> Self {
-        match self {
-            Self::Node48 {
-                num_children,
-                keys,
-                children,
-            } => {
-                let mut new_children: [Option<Box<CARTNode>>; 256] = std::array::from_fn(|_| None);
-
-                for (byte, &slot) in keys.iter().enumerate() {
-                    if slot != 255 {
-                        new_children[byte].clone_from(&children[slot as usize]);
-                    }
-                }
-
-                Self::Node256 {
-                    num_children: u16::from(*num_children),
-                    children: new_children,
-                }
-            }
-            _ => self.clone(),
         }
     }
 }
