+++ crates/velesdb-core/src/collection/graph/edge_concurrent/mod.rs
@@ -1,24 +1,24 @@
-//! Concurrent edge storage with sharding for thread-safe graph operations.
+//! Concurrent edge store with sharded locking.
 //!
 //! This module provides `ConcurrentEdgeStore`, a thread-safe wrapper around
 //! `EdgeStore` that uses sharding to reduce lock contention.
+//!
+//! Read-only queries and traversal are in `query.rs`.
 
 // SAFETY: Numeric casts in edge store sharding are intentional:
 // - u64->usize for node ID hashing: Node IDs are generated sequentially and fit in usize
 // - Used for sharding only, actual storage uses u64 for persistence
 #![allow(clippy::cast_possible_truncation)]
 
+mod query;
+
 use super::edge::{EdgeStore, GraphEdge};
 use crate::error::{Error, Result};
 use parking_lot::RwLock;
-use std::collections::{HashMap, HashSet, VecDeque};
+use std::collections::{HashMap, HashSet};
 
 /// Default number of shards for concurrent edge store.
 /// Increased from 64 to 256 for better scalability with 10M+ edges (EPIC-019 US-001).
-/// With 256 shards and 10M edges, each shard contains ~39K edges (vs ~156K with 64 shards),
-/// reducing lock contention in multi-threaded scenarios.
-///
-/// For smaller graphs, use `with_estimated_edges()` to get optimal shard count.
 const DEFAULT_NUM_SHARDS: usize = 256;
 
 /// Minimum edges per shard for efficiency.
@@ -40,25 +40,17 @@ const MAX_SHARDS: usize = 512;
 /// - **Source shard**: Full edge with outgoing + label indices (`add_edge`)
 /// - **Target shard**: Edge copy with incoming index only (`add_edge_incoming_only`)
 ///
-/// This enables O(1) lookups for both `get_outgoing(node)` and `get_incoming(node)`
-/// without cross-shard queries. Label indices are maintained only in the source shard
-/// to avoid duplication.
-///
 /// # Lock Ordering
 ///
 /// When acquiring multiple shard locks, always acquire in ascending
 /// shard index order to prevent deadlocks.
 #[repr(C, align(64))]
 pub struct ConcurrentEdgeStore {
-    shards: Vec<RwLock<EdgeStore>>,
-    num_shards: usize,
+    pub(super) shards: Vec<RwLock<EdgeStore>>,
+    pub(super) num_shards: usize,
     /// Global registry of edge IDs with source node for optimized removal.
     /// Maps edge_id -> source_node_id for O(1) shard lookup during removal.
-    ///
-    /// **FLAG-5: Changed from HashSet to HashMap intentionally.**
-    /// This trades ~8 bytes extra memory per edge for O(1) removal lookup
-    /// instead of O(shards) iteration. Worth it for graphs with frequent deletions.
-    edge_ids: RwLock<std::collections::HashMap<u64, u64>>,
+    pub(super) edge_ids: RwLock<HashMap<u64, u64>>,
 }
 
 impl ConcurrentEdgeStore {
@@ -88,27 +80,13 @@ impl ConcurrentEdgeStore {
 
     /// Creates a concurrent edge store with optimal shard count for estimated edge count.
     ///
-    /// # Shard Count Selection
-    ///
-    /// - `< 1K edges`: 1 shard (no sharding overhead)
-    /// - `1K - 64K edges`: 16-64 shards
-    /// - `64K - 1M edges`: 64-128 shards  
-    /// - `> 1M edges`: 256 shards (default)
-    ///
-    /// This avoids memory overhead of 256 shards for small graphs while
-    /// maintaining good concurrency for large graphs.
-    ///
     /// **FLAG-6: Uses integer bit manipulation for ceiling log2.**
-    /// Formula: `ceil(log2(n)) = bits - leading_zeros(n-1)` for n > 1.
-    /// This avoids floating-point imprecision from `(n as f64).log2().ceil()`.
     #[must_use]
     pub fn with_estimated_edges(estimated_edges: usize) -> Self {
         let optimal_shards = if estimated_edges < MIN_EDGES_PER_SHARD {
             1
         } else {
             let target_shards = estimated_edges / MIN_EDGES_PER_SHARD;
-            // Use integer math for log2 ceiling: avoid floating-point imprecision
-            // Formula: ceil(log2(n)) = bits - leading_zeros(n-1) for n > 1
             let power_of_2 = if target_shards <= 1 {
                 0
             } else {
@@ -121,7 +99,7 @@ impl ConcurrentEdgeStore {
 
     /// Returns the shard index for a given node ID.
     #[inline]
-    fn shard_index(&self, node_id: u64) -> usize {
+    pub(super) fn shard_index(&self, node_id: u64) -> usize {
         (node_id as usize) % self.num_shards
     }
 
@@ -152,12 +130,10 @@ impl ConcurrentEdgeStore {
         let source_shard = self.shard_index(source_id);
         let target_shard = self.shard_index(edge.target());
 
-        // Note: EdgeStore's duplicate check is now redundant but kept for safety
         if source_shard == target_shard {
             // Same shard: single lock, EdgeStore handles both indices
             let mut guard = self.shards[source_shard].write();
             guard.add_edge(edge)?;
-            // Store edge_id -> source_id for optimized removal
             ids.insert(edge_id, source_id);
         } else {
             // Different shards: acquire locks in ascending order to prevent deadlock
@@ -170,27 +146,19 @@ impl ConcurrentEdgeStore {
             let mut first_guard = self.shards[first_idx].write();
             let mut second_guard = self.shards[second_idx].write();
 
-            // Add to source shard (outgoing index)
-            // Add to target shard (incoming index)
-            // Handle errors with proper rollback
             if source_shard < target_shard {
-                // first = source, second = target
                 first_guard.add_edge_outgoing_only(edge.clone())?;
                 if let Err(e) = second_guard.add_edge_incoming_only(edge) {
-                    // Rollback first shard operation
                     first_guard.remove_edge_outgoing_only(edge_id);
                     return Err(e);
                 }
             } else {
-                // first = target, second = source
                 second_guard.add_edge_outgoing_only(edge.clone())?;
                 if let Err(e) = first_guard.add_edge_incoming_only(edge) {
-                    // Rollback second shard operation
                     second_guard.remove_edge_outgoing_only(edge_id);
                     return Err(e);
                 }
             }
-            // Insert AFTER successful shard mutations - store source for optimized removal
             ids.insert(edge_id, source_id);
         }
         Ok(())
@@ -198,31 +166,22 @@ impl ConcurrentEdgeStore {
 
     /// Removes an edge by ID using optimized 2-shard lookup.
     ///
-    /// # Performance
-    ///
-    /// Instead of iterating all 256 shards, uses stored source_id to find
-    /// the exact 2 shards (source + target) where the edge is stored.
-    ///
     /// # Concurrency Safety
     ///
     /// Lock ordering: edge_ids FIRST, then shards in ascending order.
     pub fn remove_edge(&self, edge_id: u64) {
-        // Acquire edge_ids lock FIRST (same ordering as add_edge)
         let mut ids = self.edge_ids.write();
 
-        // Get source_id for optimized shard lookup
         let Some(&source_id) = ids.get(&edge_id) else {
-            return; // Edge doesn't exist
+            return;
         };
 
-        // Get the edge to find target_id (need to read from source shard)
         let source_shard_idx = self.shard_index(source_id);
         let target_id = {
             let guard = self.shards[source_shard_idx].read();
             if let Some(edge) = guard.get_edge(edge_id) {
                 edge.target()
             } else {
-                // Edge in registry but not in shard - cleanup registry
                 ids.remove(&edge_id);
                 return;
             }
@@ -230,14 +189,9 @@ impl ConcurrentEdgeStore {
 
         let target_shard_idx = self.shard_index(target_id);
 
-        // Remove from only the relevant shards (2 max instead of 256)
         if source_shard_idx == target_shard_idx {
-            // Same shard: full removal from single shard
             self.shards[source_shard_idx].write().remove_edge(edge_id);
         } else {
-            // Cross-shard: use specialized methods for efficiency
-            // Source shard has outgoing + label indices, target shard has incoming only
-            // Acquire locks in ascending order to prevent deadlock
             let (first_idx, second_idx) = if source_shard_idx < target_shard_idx {
                 (source_shard_idx, target_shard_idx)
             } else {
@@ -246,131 +200,24 @@ impl ConcurrentEdgeStore {
             let mut first = self.shards[first_idx].write();
             let mut second = self.shards[second_idx].write();
 
-            // Use specialized removal: source gets full removal, target gets incoming-only
             if source_shard_idx < target_shard_idx {
-                first.remove_edge(edge_id); // Source: full cleanup
-                second.remove_edge_incoming_only(edge_id); // Target: incoming index only
+                first.remove_edge(edge_id);
+                second.remove_edge_incoming_only(edge_id);
             } else {
-                first.remove_edge_incoming_only(edge_id); // Target: incoming index only
-                second.remove_edge(edge_id); // Source: full cleanup
+                first.remove_edge_incoming_only(edge_id);
+                second.remove_edge(edge_id);
             }
         }
 
-        // Remove from global registry
         ids.remove(&edge_id);
-    }
-
-    /// Gets all outgoing edges from a node (thread-safe).
-    #[must_use]
-    pub fn get_outgoing(&self, node_id: u64) -> Vec<GraphEdge> {
-        let shard = &self.shards[self.shard_index(node_id)];
-        let guard = shard.read();
-        guard.get_outgoing(node_id).into_iter().cloned().collect()
-    }
-
-    /// Gets all incoming edges to a node (thread-safe).
-    #[must_use]
-    pub fn get_incoming(&self, node_id: u64) -> Vec<GraphEdge> {
-        let shard = &self.shards[self.shard_index(node_id)];
-        let guard = shard.read();
-        guard.get_incoming(node_id).into_iter().cloned().collect()
-    }
-
-    /// Gets neighbors (target nodes) of a given node.
-    #[must_use]
-    pub fn get_neighbors(&self, node_id: u64) -> Vec<u64> {
-        self.get_outgoing(node_id)
-            .iter()
-            .map(GraphEdge::target)
-            .collect()
-    }
-
-    /// Gets outgoing edges filtered by label (thread-safe).
-    ///
-    /// # Performance Note
-    ///
-    /// This method delegates to the underlying `EdgeStore::get_outgoing_by_label`
-    /// which uses the composite index `(source_id, label) -> edge_ids` for O(1) lookup
-    /// when available (EPIC-019 US-003). Falls back to filtering if index not populated.
-    #[must_use]
-    pub fn get_outgoing_by_label(&self, node_id: u64, label: &str) -> Vec<GraphEdge> {
-        let shard_idx = self.shard_index(node_id);
-        let shard = self.shards[shard_idx].read();
-        shard
-            .get_outgoing_by_label(node_id, label)
-            .into_iter()
-            .cloned()
-            .collect()
-    }
-
-    /// Gets incoming edges filtered by label (thread-safe).
-    #[must_use]
-    pub fn get_incoming_by_label(&self, node_id: u64, label: &str) -> Vec<GraphEdge> {
-        self.get_incoming(node_id)
-            .into_iter()
-            .filter(|e| e.label() == label)
-            .collect()
-    }
-
-    /// Gets all edges with a specific label across all shards.
-    ///
-    /// # Performance Warning
-    ///
-    /// This method iterates through ALL shards and aggregates results.
-    /// For large graphs with many shards, this can be expensive.
-    /// Consider using `get_outgoing_by_label(node_id, label)` if you know
-    /// the source node, which is O(k) instead of O(shards × edges_per_label).
-    ///
-    /// # Use Cases
-    ///
-    /// - Schema introspection (listing all edges of a type)
-    /// - Batch operations on all edges of a type
-    /// - Testing and debugging
-    #[must_use]
-    pub fn get_edges_by_label(&self, label: &str) -> Vec<GraphEdge> {
-        self.shards
-            .iter()
-            .flat_map(|shard| {
-                shard
-                    .read()
-                    .get_edges_by_label(label)
-                    .into_iter()
-                    .cloned()
-                    .collect::<Vec<_>>()
-            })
-            .collect()
-    }
-
-    /// Checks if an edge with the given ID exists.
-    #[must_use]
-    pub fn contains_edge(&self, edge_id: u64) -> bool {
-        self.edge_ids.read().contains_key(&edge_id)
-    }
-
-    /// Gets an edge by ID using optimized source shard lookup.
-    ///
-    /// Returns `None` if the edge doesn't exist.
-    #[must_use]
-    pub fn get_edge(&self, edge_id: u64) -> Option<GraphEdge> {
-        // Get source_id from registry for direct shard lookup
-        let source_id = *self.edge_ids.read().get(&edge_id)?;
-        let shard_idx = self.shard_index(source_id);
-        self.shards[shard_idx].read().get_edge(edge_id).cloned()
     }
 
     /// Removes all edges connected to a node (cascade delete, thread-safe).
     ///
-    /// Handles cross-shard cleanup: collects all edges, then removes from all
-    /// relevant shards with proper lock ordering to prevent deadlocks.
-    ///
     /// # Concurrency Safety
     ///
     /// Lock ordering: edge_ids FIRST, then shards in ascending order.
-    /// This matches add_edge/remove_edge ordering to prevent deadlocks.
-    /// The edge_ids lock is held throughout to prevent add_edge from
-    /// inserting IDs we're about to remove.
     pub fn remove_node_edges(&self, node_id: u64) {
-        // CRITICAL: Acquire edge_ids lock FIRST (same ordering as add_edge/remove_edge)
         let mut ids = self.edge_ids.write();
 
         let node_shard = self.shard_index(node_id);
@@ -391,7 +238,7 @@ impl ConcurrentEdgeStore {
             (outgoing, incoming)
         };
 
-        // Phase 2: Collect all shards that need cleanup
+        // Phase 2: Collect all shards that need cleanup (BTreeSet = sorted ascending)
         let mut shards_to_clean: std::collections::BTreeSet<usize> =
             std::collections::BTreeSet::new();
         shards_to_clean.insert(node_shard);
@@ -404,7 +251,6 @@ impl ConcurrentEdgeStore {
         }
 
         // Phase 3: Acquire shard locks in ascending order and perform cleanup
-        // BTreeSet iteration is already sorted ascending
         let mut guards: Vec<_> = shards_to_clean
             .iter()
             .map(|&idx| (idx, self.shards[idx].write()))
@@ -413,27 +259,22 @@ impl ConcurrentEdgeStore {
         // Phase 4: Clean up edges in all shards
         for (shard_idx, guard) in &mut guards {
             if *shard_idx == node_shard {
-                // Main shard: full cleanup
                 guard.remove_node_edges(node_id);
             } else {
-                // Other shards: clean only the cross-shard edge entries
                 for (edge_id, target) in &outgoing_edges {
                     if self.shard_index(*target) == *shard_idx {
-                        // This edge's incoming index is in this shard
                         guard.remove_edge_incoming_only(*edge_id);
                     }
                 }
                 for (edge_id, source) in &incoming_edges {
                     if self.shard_index(*source) == *shard_idx {
-                        // This edge's outgoing index is in this shard
                         guard.remove_edge_outgoing_only(*edge_id);
                     }
                 }
             }
         }
 
-        // Phase 5: Remove edge IDs from global registry (still holding lock)
-        // Note: Use a set to deduplicate IDs (self-loops appear in both lists)
+        // Phase 5: Remove edge IDs from global registry
         let mut removed: HashSet<u64> = HashSet::new();
         for (edge_id, _) in &outgoing_edges {
             if removed.insert(*edge_id) {
@@ -445,54 +286,6 @@ impl ConcurrentEdgeStore {
                 ids.remove(edge_id);
             }
         }
-    }
-
-    /// Traverses the graph using BFS from a starting node.
-    ///
-    /// Returns all nodes reachable within `max_depth` hops.
-    ///
-    /// Uses Read-Copy-Drop pattern to avoid holding locks during traversal.
-    #[must_use]
-    pub fn traverse_bfs(&self, start: u64, max_depth: u32) -> Vec<u64> {
-        let mut visited = HashSet::new();
-        let mut queue = VecDeque::new();
-        queue.push_back((start, 0u32));
-
-        while let Some((node, depth)) = queue.pop_front() {
-            if depth > max_depth || !visited.insert(node) {
-                continue;
-            }
-
-            // Read-Copy-Drop pattern: copy neighbors and drop guard immediately
-            let neighbors: Vec<u64> = {
-                let shard = &self.shards[self.shard_index(node)];
-                let guard = shard.read();
-                guard
-                    .get_outgoing(node)
-                    .iter()
-                    .map(|e| e.target())
-                    .collect()
-            }; // Guard dropped here
-
-            for neighbor in neighbors {
-                if !visited.contains(&neighbor) {
-                    queue.push_back((neighbor, depth + 1));
-                }
-            }
-        }
-
-        visited.into_iter().collect()
-    }
-
-    /// Returns the total edge count across all shards.
-    ///
-    /// Uses outgoing edge count to avoid double-counting edges that span shards.
-    #[must_use]
-    pub fn edge_count(&self) -> usize {
-        self.shards
-            .iter()
-            .map(|s| s.read().outgoing_edge_count())
-            .sum()
     }
 }
