+++ crates/velesdb-core/src/database/database_tests.rs
@@ -1,6 +1,6 @@
 use super::*;
-use super::*;
 use crate::collection::graph::GraphEdge;
+use crate::point::Point;
 use crate::velesql::Parser;
 use tempfile::tempdir;
